Start refactoring DumpDataCollector
